Sanitizer method moved to BundleUtils Class + removed main
